feat: implement operator type recognition in new operators.c file
