+++ src/lexer/lexer_utils.c
@@ -6,7 +6,7 @@
 /*   By: nbuchhol <nbuchhol@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/01 11:59:48 by nbuchhol          #+#    #+#             */
-/*   Updated: 2025/06/02 17:33:38 by nbuchhol         ###   ########.fr       */
+/*   Updated: 2025/06/02 20:48:31 by nbuchhol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,25 +43,6 @@ int	is_operator(char c)
 int	is_metacharacter(char c)
 {
 	return (ft_isspace(c) || is_operator(c));
-}
-
-/**
- * @brief Get token type based on string content.
- * @param char* Pointer to current position in string.
- * @param int index of the current position in string.
- * @return Corresponding token type enum.
- */
-t_token_type	get_operator_type(char *input, int pos)
-{
-	if (input[pos + 1] && input[pos] == '>' && input[pos + 1] == '>')
-		return (TOKEN_REDIR_APPEND);
-	if (input[pos] == '|')
-		return (TOKEN_PIPE);
-	if (input[pos] == '<')
-		return (TOKEN_REDIR_IN);
-	if (input[pos] == '>')
-		return (TOKEN_REDIR_OUT);
-	return (TOKEN_WORD);
 }
 
 /**
